backend/pyproject.toml
[project]
name = "vocab-backend"
version = "0.1.0"
description = "FastAPI backend for the personal vocabulary app."
requires-python = ">=3.12,<3.15"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.0",
    "psycopg[binary]>=3.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
